test(Answer): Answer 객체에서 담당했던 랜덤 값에 대한 테스트를 RandomNumberGenerator로 이동
Answer 객체는 생성된 리스트를 받아 생성되는 객체일 뿐입니다. 따라서, RandomNumber에 대한 검증은 RandomNumberGenerator가 해야한다고 판단하여 테스트 로직을 이동했습니다,
