+++ src/test/java/baseball/domain/baseball_number/BaseballNumberTest.java
@@ -14,11 +14,7 @@ class BaseballNumberTest {
     @Test
     void 정답_숫자_리스트_안에_ball의_조건을_만족하는_SingleNumber가_있을_경우_1을_반환한다() {
         BaseballNumber baseballNumber = new BaseballNumber(0, 1);
-        List<BaseballNumber> answer = List.of(
-                new BaseballNumber(0, 2),
-                new BaseballNumber(1, 1),
-                new BaseballNumber(2, 3)
-        );
+        List<BaseballNumber> answer = BaseballNumberListGenerator.baseballNumberList("213");
 
         assertThat(baseballNumber.equalsOnlyNumber(answer)).isTrue();
     }
@@ -26,11 +22,7 @@ class BaseballNumberTest {
     @Test
     void 정답_숫자_리스트_안에_ball의_조건을_만족하는_SingleNumber가_없을_경우_0을_반환한다() {
         BaseballNumber baseballNumber = new BaseballNumber(0, 1);
-        List<BaseballNumber> answer = List.of(
-                new BaseballNumber(0, 2),
-                new BaseballNumber(1, 4),
-                new BaseballNumber(2, 3)
-        );
+        List<BaseballNumber> answer = BaseballNumberListGenerator.baseballNumberList("243");
 
         assertThat(baseballNumber.equalsOnlyNumber(answer)).isFalse();
     }
